spark-parser: Insert waiting and scheduling driver activities
This commit implements the logic of the `spark-driver-hack` flag of
`pag-construction` as part of the Spark preprocessor. Basically, we
assume that the driver was `Scheduling` if the unknown activity
ends in a submitted task, and we assume that the driver was
`Waiting` before it receives the task result message.

The resulting PAG is slightly different at window boundaries
from what was implemented in `pag-construction`: The logic implemented
in this commit runs before slicing, whereas the old
heuristics was applied after slicing.

This commit adds some additional inline comments as well.

Signed-off-by: Sebastian Wicki <swicki@inf.ethz.ch>